requirements.txt
Django==1.8
django-allauth==0.24.1
django-grappelli==2.8.1
mongoengine==0.10.6
oauthlib==1.0.3
psycopg2==2.6.1
pymongo==3.2.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
